Update GUI layouts
Move 'Store position' button to main dialog in 'Current camera' section. When the 'Store position' button was in the 'Camera settings' dialog it implied that it was committing the changes made in that dialog, which is simply not true.
Move 'Copy camera' button to 'Current camera' section because copy button is copies current camera.
Change text field to combobox for camera category setting